projects/03_coffee_shop_full_stack/starter_code/backend/src/requirements.txt
alembic==1.4.2
appdirs==1.4.4
autopep8==1.5.3
Babel==2.8.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.5
pycodestyle==2.6.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.12.0
SQLAlchemy==1.3.18
toml==0.10.1
virtualenv==20.0.26
Werkzeug==1.0.1
wrapt==1.11.1
WTForms==2.3.1
python-jose==3.1.0

